Create README.md
During my co-op as a customer service representative at Broil King, I realized that there were a lot of straightforward questions that could easily be answered and that customers would call not having the information necessary to be assisted. The easiest way this could be remediated was to create a customer service bot that could recognize their questions and give them an immediate answer. I followed Federico Cotogno's text recognition chat bot.