Update README.md
This Bot was created following a tutorial by federicocotogno. The user supplies a word or sentence, which is then analyzed for required or recognized words. Based on the number of recognized words the user supplies, a percentage of recognized words over the total recognized words will be generated. The response with the highest percentage will be the one used to reply to the user. The required word responses need the required word to be used to be prompted. This helps make the response more specific. There is also a long repsonse section for lenghtier answers to avoid making the code messy. Lastly, we have a def_unknown section for words that are not part of the required words or the recognized words. This let's the customer know the bot will not be able to help them instead of generating an error.